Add is_accomplished feature

+++ crowdfunding/views.py
@@ -25,9 +25,13 @@ def create_crowdfund(request):
     if request.method == "POST":
         form = CrowdfundForm(request.POST)
         if form.is_valid():
+            received = request.POST["received"]
+            target = request.POST["target"]
             new_crowdfund = form.save(commit=False)
             new_crowdfund.user = Profile.objects.get(user=request.user)
             new_crowdfund.user.add_poin(CREATE_POINT)
+            if received == target:
+                new_crowdfund.is_accomplished = True
             new_crowdfund.save()
             return redirect("crowdfunding:show_crowdfundings")
 
@@ -38,13 +42,22 @@ def create_crowdfund(request):
 @login_required(login_url="/login/")
 def edit_crowdfund(request, id):
     crowdfund = Crowdfund.objects.get(pk=id)
+    if crowdfund.user.pk != request.user.pk:
+        return redirect("crowdfunding:show_crowdfundings")
 
     form = CrowdfundForm(instance=crowdfund)
 
     if request.method == "POST":
         form = CrowdfundForm(request.POST, instance=crowdfund)
         if form.is_valid():
-            form.save()
+            received = request.POST["received"]
+            target = request.POST["target"]
+            edited_crowdfund = form.save(commit=False)
+            if received == target:
+                edited_crowdfund.is_accomplished = True
+            else:
+                edited_crowdfund.is_accomplished = False
+            edited_crowdfund.save()
             return redirect("crowdfunding:show_crowdfundings")
 
     context = {"form": form}
@@ -54,6 +67,9 @@ def edit_crowdfund(request, id):
 @login_required(login_url="/login/")
 def delete_crowdfund(request, id):
     crowdfund = Crowdfund.objects.get(pk=id)
+    if crowdfund.user.pk != request.user.pk:
+        return redirect("crowdfunding:show_crowdfundings")
+
     crowdfund.delete()
     return redirect("crowdfunding:show_crowdfundings")
